+++ packages/sdk-js/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## **1.0.0** - May 1, 2024
 
-- New `init` function as a replacement for `loadFeatures`.
-- Support payload import/export via `init({ payload })`, `setPayload`, and `getPayload`. Makes it easier to share data between front-end, back-end, and other services.
+- New `init` and `initSync` functions as a replacement for `loadFeatures`.
+- Support payload import/export via `init({ payload })`, `setPayload`, `getPayload`, and `getDecryptedPayload`. Makes it easier to share data between front-end, back-end, and other services.
 - Ability to block or control experiments by various context properties
   - `blockedChangeIds` targets individual experiments
   - `disableVisualExperiments` prevents visual experiments from running
@@ -12,6 +12,7 @@
   - `disableCrossOriginUrlRedirectExperiments` blocks redirects if the origin changes
   - `disableExperimentsOnLoad` prevents AutoExperiments from running automatically; `triggerAutoExperiments` manually triggers them.
 - Provide a custom DOM mutation method by setting context `applyDomChangesCallback`. Useful for server-side rendering visual experiments.
+- Fix bug when passing in `stickyBucketAssignmentDocs` to the GrowthBook constructor
 - New `jsInjectionNonce` setting to add a nonce onto any injected <script\> tags. This provides a safer alternative to allowing `unsafe-inline` in your Content Security Policy.
 - Get a list of triggered experiments by their `changeId` (a new, more-specific AutoExperiment identifier) via `getCompletedChangeIds`.
 - Disable in-memory cache via `configureCache({ disableLocalCache: true })` or the new GrowthBook constructor option `disableCache`.